refactor: 移动isGroupManager方法到合适的地方
