+++ src/shared/helper/group.ts
@@ -1,7 +1,27 @@
 import { GroupInfo } from '@redux/types/group';
-import { isGroupManager } from '@shared/model/group';
 import { GroupPanel } from '@shared/types/panel';
 import _isNil from 'lodash/isNil';
+import _isArray from 'lodash/isArray';
+
+/**
+ * TODO: 这个文件应当移动到 `/src/shared/helper/group.ts` 中
+ * 判断某个用户是否为团管理员
+ * @param groupInfo 团信息
+ */
+export function isGroupManager(
+  groupInfo: GroupInfo,
+  userUUID: string
+): boolean {
+  if (_isNil(groupInfo) || _isNil(userUUID)) {
+    return false;
+  }
+
+  return (
+    groupInfo.owner_uuid === userUUID ||
+    (_isArray(groupInfo.managers_uuid) &&
+      groupInfo.managers_uuid.includes(userUUID))
+  );
+}
 
 /**
  * 判断团面板是否可见
